✨ feat: route VM suspend/resume through the park gate (Phase B PR2)
Compose the non-terminal suspend triggers on one reason set (ADR-017,
#682) so view-hide, app-background, and user-pause can't desync the
shared SuspendController:

- VM gains a weak `session` back-ref (set in startGuarded) and routes
  pauseSimulation/resumeSimulation (.userPause) and the scene-phase
  background/foreground suspend+resume (.appBackground) through
  `routePark/routeUnpark` → the session gate. Falls back to the
  controller directly when no session owns the run (fixture tests),
  preserving pre-Phase-B behaviour.
- The terminal trio (cancelSimulation, run()/resume() defers) stays
  outside the gate — over-resume is harmless as a run ends.
- handleBackgroundActivation honours the view-hide precedence: a
  parked-away run releases the BG task and stays parked in memory rather
  than switching to off-screen CPU inference (Variant 3).

Routing (gate vs fallback) is unit-tested deterministically via
routePark/routeUnpark; the away-case precedence by an in-flight
integration test. SimulationSessionTests split into a sibling extension
file to stay under file_length (testing.md).
