+++ Pastura/Pastura/App/SimulationSession.swift
@@ -177,6 +177,10 @@ final class SimulationSession {
     self.tab = tab
     parkReasons = []
     let task = Task { await body(viewModel) }
+    // Back-reference (weak on the VM) so the VM's non-terminal suspend/resume
+    // (user-pause, scene-phase) route through this session's park gate instead
+    // of touching the SuspendController directly — see `routePark(reason:)`.
+    viewModel.session = self
     // Intentional dual assignment of the SAME task: `driveTask` is the
     // session's cancellation handle (``end()``), and `viewModel.runTask` is the
     // VM's (`cancelSimulation()` — user cancel, memory warning). Two independent
